refactor(agent_backends): extract codex_gate submodule
Move the alternative-backends gate, native-Codex gate, and one-shot
legacy-id migration cluster out of the agent_backends facade and
into a focused codex_gate submodule. Symbols moved:

- Constants: ALTERNATIVE_BACKENDS_SETTING_KEY,
  NATIVE_CODEX_BACKEND_ID, LEGACY_NATIVE_CODEX_BACKEND_ID,
  LEGACY_CODEX_SUBSCRIPTION_BACKEND_ID, NATIVE_CODEX_SETTING_KEY,
  LEGACY_NATIVE_CODEX_SETTING_KEY, FIRST_CLASS_BACKENDS_PROMOTION_KEY.
- Predicates: is_always_on_alt_backend, is_codex_gate_backend_id,
  codex_backend_hidden_by_gate, is_legacy_codex_backend_id.
- DB readers / writers: native_codex_enabled,
  alternative_backends_enabled, promote_first_class_backend_gates,
  migrate_legacy_codex_backend_settings.
- Guards: ensure_native_codex_enabled,
  ensure_backend_id_allowed_by_gate, ensure_backend_allowed_by_gate.
- Default-list builder: default_backends_for_gate.

Five tests covering gate-only behavior move alongside their helpers
(alternative_backends_promote_on_by_default,
saved_false_backend_gates_are_flipped_during_promotion,
native_codex_command_guard_requires_models_gate_after_promotion,
alternative_backends_can_still_be_disabled_without_codex,
legacy_codex_backend_settings_migrate_to_canonical_codex).
Cross-cutting tests (load/save round-trips with hidden Codex passthrough)
stay with their primary subject — they will follow when the config
submodule lands.

Cross-module visibility uses `pub(super)` so nothing leaks beyond the
agent_backends module. The 9 #[tauri::command] entry points, the
disabled-twin contract, and every wire shape are unchanged.

cargo fmt --check, cargo clippy -Dwarnings on the three CI crates,
and the agent_backends test suite all pass (111 passed; test count
matches baseline).
